Moved commands to package
